collect_rent.php
Using php to create a webform. This form will talk to the Pay_Rent sol and collect rent from your tenants. As eth for v1. This page is way too complicated I left a lot in there to give a more professional look. I'm also making a simple version of this, logo and form. 
Little to No formatting.

New trend is coming also, this new trend will allow you enter your wallet info and simplified this process for the user.,